+++ lib/lan_scanner.dart
@@ -1,5 +1,6 @@
 library lan_scanner;
 
 export 'package:lan_scanner/src/lan_scanner.dart';
-export 'package:lan_scanner/src/models/device_address.dart';
+export 'package:lan_scanner/src/models/device_model.dart';
+export 'package:lan_scanner/src/models/progress_model.dart';
 export 'package:lan_scanner/src/utils/ip_to_subnet.dart';