+++ lib/src/models/progress_callback.dart
@@ -1,2 +1,5 @@
+// Project imports:
+import 'package:lan_scanner/src/models/progress_model.dart';
+
 /// Typedef for progress callbacks
-typedef ProgressCallback = void Function(String progress);
+typedef ProgressCallback = void Function(ProgressModel progress);