+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.0.0
+
+* Changed name of DeviceAddress to DeviceModel
+* ProgressCallback now return a ProgressModel instead of a String
+* preciseScan() now uses StreamController instead of generator function
+* Fix typos
+* Bug fixes
+* Updated docs
+* Updated example
+
 ## 1.0.3
 
 * Added lint rules